[Refactor] Move and restructure FAQ components
