+++ src/pages/Faq/Children/Roadmap/Roadmap.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Header } from "@/components/sections/Header";
-import RoadmapFeat from "./RoadmapFeat";
-import RoadmapChart from "./RoadmapChart";
+import RoadmapFeat from "@/components/layouts/Faq/Children/Roadmap/RoadmapFeat";
+import RoadmapChart from "@/components/layouts/Faq/Children/Roadmap/RoadmapChart";
 import { Footer } from "@/components/sections/Footer";
 
 export default function Roadmap() {
